+++ src/sdmf/exception/ExtractionException.py
@@ -1,9 +1,9 @@
-from sdmf.exception.BasePipelineException import BasePipelineException
+from sdmf.exception.BaseException import BaseException
 
-class ExtractionException(BasePipelineException):
+class ExtractionException(BaseException):
     def __init__(self, message=None, details=None, original_exception=None):
         super().__init__(
-            message or "Extraction Error",
+            message or "Extraction Exception",
             details=details,
             original_exception=original_exception
         )