+++ src/sdmf/exception/ValidationError.py
@@ -1,9 +1,9 @@
-from sdmf.exception.BasePipelineException import BasePipelineException
+from sdmf.exception.BaseException import BaseException
 
-class ValidationError(BasePipelineException):
+class ValidationError(BaseException):
     def __init__(self, message=None, details=None, original_exception=None):
         super().__init__(
-            message or "Extraction Error",
+            message or "Validation Error",
             details=details,
             original_exception=original_exception
         )